requirements.txt
accelerate==0.34.2
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
blobfile==3.0.0
bs4==0.0.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
fastapi==0.115.0
filelock==3.16.1
fire==0.6.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
llama_cpp_python==0.3.0
llama_models==0.0.36
llama_stack==0.0.36
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psutil==6.0.0
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.1.3
starlette==0.38.6
starlette-context==0.3.6
sympy==1.13.3
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
wcwidth==0.2.13
